Computer World - API Connection & DIsplay
I have managed to update both the API & the App, they can now both communicate with each other. I implemented CORS into the API which allows the front end to communicate with the API. I am able to receive the JSON data from the api and display it.

I will find a way to manipulate this data after designing a layout for the site.