Computer World - Data Display Overhaul
Data is now being fully displayed in tables. I applied bootstrap to the project to make it look a lot cleaner than before. I need to add a scroll effect to the table to prevent the data from being display all over the page. When a user selects a row in the table. That data is then displayed in the form below it. Buttons do not work at the moment. Next step is to get the CRUD features working.