Computer World - Fully Implemented CRUD
The front-end is now fully functional. Information is displayed and can be selected. Once selected, the values are displayed inside the form. The values can be changed and the changes can be submitted. The record is updated in the database. If the user selects a row, they can simply press the delete button and it will remove them from the database. If no index is selected, they can fill the form with data and then press the create button which will generate a new record. The user interface was also changed  thanks to Bootstrap. The information is easy to read.

Plenty of problems are present in the system.
 - Form validation is required.
 - Feedback to the user is needed.
 - Changes to the API to prevent wrongful inputs.
 - Refactoring of the code to prevent duplicate inputs as much as possible
 - User Interface needs to be finished.