Computer World - Implemented Feedback Part 3
Implemented the Delete function into the Table components and List View Files.

Implemented the Create function into the Create View files. Imported some of my old code to get it to function.

View pages have been set, when a user creates a new record, it routes them to the page to view the details, same thing happens when clicking the view button in the table.

Need to insert an update button which sends them to an editing page where the data is  auto-filled into a form based on the id.

Made some UI tweaks to make it easier to read for now, will finish UI last.

Need to implement the Update, error handling, and validation into the code and the project should be complete.