Computer World - More Feedback Implementation
Spoke with Cameron yesterday and he gave me a little feedback for my application. I have managed to implement all of them besides the responsive design just yet. I am working on that now.

Implemented toasts. Errors are no longer swallowed and are instead caught and displayed.

Relocated the window from the service file into the components files. Implemented a delete handler which the window is stored in and checks to see if the user confirms to delete the selected record or not.

Replaced any and all a tags with RouterLink