Computer World - Responsiveness and Error Handling
The website is much more responsive now and doesn't lock to the sidebar. It instead shows the centre of the page. The side bar can now be closed and re-opened by hitting a little arrow.

I have implemented error handling and validation to all of the create and updates views. The user can no longer input the incorrect variables. Need to restrict the amount of characters or numbers they can input.

Reduced duplication of the schema by implementing it into its own typescript file.