Computer World - Responsive Design
I have implemented some new css which allows all the tables on the application to become responsive and informative on mobile devices. The application can now be used both on desktop/mobile. I believe the application is now fully responsive. It still needs some UI tweaks to make it look nice, but it is fully functional now