Computer World - Feedback Implementation
Implemented all of the feedback I was given. Besides an error that Cameron was getting when trying to run the project. Services no longer swallow errors. and the code has been massively reduced. The try/catches have been added to the views. Returning toast errors. I decided to keep the schemas in their own file as they are used in multiple places. Moved the components into their own folders to make them easier to read.